documentation_creator/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.6
Deprecated==1.2.14
distro==1.9.0
docx==0.2.4
docx2python==2.10.1
exceptiongroup==1.2.1
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.56
lxml==5.2.1
marshmallow==3.21.2
multidict==6.0.5
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.28.1
orjson==3.10.3
packaging==23.2
pandas==2.0.3
paragraphs==0.2.1
pillow==10.3.0
pinecone-client==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pydeck==0.9.1
PyDrive2==1.19.0
PyGithub==2.3.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
streamlit==1.34.0
tenacity==8.3.0
tiktoken==0.6.0
toolz==0.12.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
wrapt==1.16.0
yarl==1.9.4
